ticket-help/requirements.txt
discord.py>=2.3.2

firebase-admin>=6.4.0
google-cloud-firestore>=2.14.0

python-dotenv>=1.0.0

urllib3<2.0
